requirements.txt
# Core Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
psutil>=5.9.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
greenlet==3.0.3

# CLI and Terminal UI
click>=8.0.0
prompt-toolkit>=3.0.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1

# Development
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Logging and Monitoring
python-json-logger>=2.0.7

# Additional dependencies
python-dotenv>=1.0.0

websockets==12.0
aiohttp>=3.9.0
cryptography==42.0.5
PyJWT==2.8.0
aioredis==2.0.1
pyyaml==6.0.1

structlog==23.2.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

python-dateutil>=2.8.2
networkx>=3.2.1
asyncio>=3.4.3
sqlalchemy>=2.0.23
sqlalchemy-utils>=0.41.1
alembic>=1.12.1
psycopg2-binary>=2.9.9
redis>=5.0.1
prometheus-client>=0.19.0
pydantic-settings>=2.1.0
